// admin-dashboard/src/app/services/models/consignor-profile-view.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ConsignorProfileView {
  businessName?: string;
  createdAt?: string;
  firstName?: string;
  lastName?: string;
  phone?: string;
  publicId?: string;
  remark?: string;
  status?: string;
  tradeLicence?: string;
  updatedAt?: string;
}