// admin-dashboard/src/app/services/models/page-response-consignor-profile-view.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ConsignorProfileView } from '../models/consignor-profile-view';
export interface PageResponseConsignorProfileView {
  content?: Array<ConsignorProfileView>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}